refactor: rename helper
(cherry picked from commit 296389741a423a55130bce5a02c540d1d1b1a2da)
