+++ src/assertPathWithType.js
