+++ src/commands/compress.command.js
@@ -2,7 +2,7 @@ import p from "node:path";
 import stream from "node:stream/promises";
 import fs from "node:fs";
 import z from "node:zlib";
-import { assertPathWithType } from "../isExist.js";
+import { assertPathWithType } from "../assertPathWithType.js";
 
 export const compressCommand = {
   name: "compress",
